+++ lib/internal/codec/lc_decoder.dart
