+++ lib/internal/codec/lc_encoder.dart
